Make sure block store is empty for running new node
